[Refactor] RefreshToken Converter 변경
